Use counter-type metrics for replications
We add a new Prometheus metric, a counter-type metric that can
have its value set directly. This is in contrast to the Go
prometheus client's Counter, which can only be incremented (and
is heavily optimised to that use-case).

We do this because the metrics we get from CouchDB are already
counters. With the prometheus client builtins, we can only use
a Gauge if we want to set the value directly. However, the
gauge-type metrics don't offer the same affordances in the Prometheus
UX --- for example, a counter-type metric can more easily be
worked into a rate over time graph, and rollovers to zero (eg on
CouchDB restarts) are better handled.

Therefore we want to emit as a counter-type but be able to directly
set the counter value from the counter we're getting from CouchDB.
The most natural way to do that I could see was to create a simple
custom metric type that allowed setting the value. It's pretty
simple in the end because we want to do very little and we don't
care about performance very much, alongside the fact that we can
leverage a lot of functionality that already exists in the prometheus
client library, like MetricVec to manage all the metrics we create
via dynamic labels.