Use counters in active tasks monitor where more appropriate
This commit updates the two "changed processed/done" metrics to
use the SettableCounter metric. The total changes metrics are
left as gauges as that feels more correct for those.

Given the confusing naming of the metrics from CouchDB, when
placed into a Prometheus context, the Prometheus metrics are
given names more in keeping with Prometheus conventions. The
variables are also made explicitly *Gauge or *Counter to remove
the ambiguity stemming from the CouchDB/Prometheus mismatch.
It's ugly but I think it brings clarity.